feat(auth): signup polish, success screen, shared CTA button
- add signup success route/screen; terms & privacy SVG assets
- signup step 2: digits-only experience input, decimal hourly rate
  with $ prefix; Step2Payload.hourlyRate int -> double
- extract AppCtaButton and adopt across auth, profile, availability,
  and onboarding screens
- re-enable LoggingInterceptor in debug builds
- file manager design PDF; migration log and handoff notes

+++ lib/features/auth/domain/repositories/auth_repository.dart
@@ -46,7 +46,7 @@ class Step2Payload {
   final int crewMemberId;
   final List<int> primaryRoleIds;
   final int yearsOfExperience;
-  final int hourlyRate;
+  final double hourlyRate;
   final String bio;
   final List<int> skillIds;
   final List<int> equipmentIds;
